Move dataDimensionType property to DimensionalObject (#3190)
Move dataDimensionType property to DimensionalObject

(cherry picked from commit 9327ce13ec04f829b1791887b621d78d41e4b9dc)
