+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/dataset/DataSetService.java
@@ -36,7 +36,6 @@ import org.hisp.dhis.period.Period;
 import org.hisp.dhis.period.PeriodType;
 import org.hisp.dhis.user.User;
 
-import java.util.Collection;
 import java.util.Date;
 import java.util.List;
